add license to setup.py (GPLv3 acc. to github project) add tools subpackage move opviewer to evmlab/tools update __main__ to make python -m evmlab opviewer <args> available create stub opviewer.py in utilities Output DecodingError if we fail to decode the inputstring
