+++ evmlab/tools/__init__.py
@@ -1,0 +1,6 @@
+"""
+Move reusable utility code that can be made available via cmdline at python -m evmlab <tool> as a module to this package
+
+example:
+    see opviewer
+"""
