+++ setup.py
@@ -9,10 +9,12 @@ setup(name='Evmlab',
       description='Ethereum EVM utilities',
       author='Martin Holst Swende',
       author_email='martin.swende@ethereum.org',
+      license="GPLv3",
       keywords=["ethereum", "transaction", "debugger"],
       url="https://github.com/holiman/evmlab/",
       download_url="https://github.com/holiman/evmlab/tarball/v%s"%version,
-      packages=['evmlab'],
+      packages=['evmlab',
+                'evmlab.tools'],
       install_requires=["requests", "web3", "eth-hash[pycryptodome]", "rlp>=1.0"],
       extras_require={"consolegui": ["urwid"],
                       "abidecoder": ["ethereum-input-decoder"],
